Add files via upload
This is just a snapshot of a quick analysis I did that looked at caviar sales from the prior year. The blue bars show the units sold, and the red dots show the margin dollars associated with each unit sold. 
The goal was to help simplify the inventory management process: we only had so many slots for caviars this year, which options that we've carried were worth bringing back in.  The shelf-life of these caviars was usually 60 to 90 days, so including February was important, since there can bee a second selling opportunity for Valentines Day if the product didn't sell out for New Year's. 

PS
The reason the sales data for NYE is marked in January is because the data was based on invoiced product, not ordered product, so invoices were processed after the Holiday.