Add files via upload
A workbook that calculated final pricing for product packaged at our distribution center vs raw product, allowing for the account manager to easily navigate the different pricing options based on hypothetical product costs, packaging costs, and labor costs, as well as production totals. This had to be extremely simple to use and edit, that's why there is a set of calculations list boxes, and a separate page for talking through the data with the account manager and client.